Search for "storyPaths" method in superclasses
Previously "storyPaths" method was searched in all methods of the test
class (with any visibility modifier) and in public methods of test class
superclasses. Now all methods of superclasses are included in the search.
It may be useful for cases when there is some common class with protected
"storyPaths" method and there is a set of sub-classes adding custom logic,
but not overriding "storyPaths" method.